search/requirements.txt
grpcio
grpcio-tools
motor
beanie
loguru
PyYAML